+++ packages/components/bolt-card-replacement/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.2](https://github.com/boltdesignsystem/bolt/compare/v2.15.1...v2.15.2) (2020-01-21)
+
+**Note:** Version bump only for package @bolt/components-card-replacement
+
+
+
+
+
+# [2.15.0](https://github.com/boltdesignsystem/bolt/compare/v2.14.3...v2.15.0) (2020-01-17)
+
+
+### Bug Fixes
+
+* linting ([a5734fe](https://github.com/boltdesignsystem/bolt/commit/a5734fe047d59db4274392f782b96a0a5c42c6b1))
+* update Jest snapshots ([e6bf37b](https://github.com/boltdesignsystem/bolt/commit/e6bf37b71dba8b50356e36cfd5b835ccf627d79d))
+
+
+### Features
+
+* add blocks to card-replacement to allow override in academy ([565a17f](https://github.com/boltdesignsystem/bolt/commit/565a17fae5ca4daddcf221319abf8717dc0fb386))
+
+
+
+
+
+## [2.14.3](https://github.com/boltdesignsystem/bolt/compare/v2.14.2...v2.14.3) (2020-01-16)
+
+
+### Bug Fixes
+
+* fix HTML rendering / escaping issue in Drupal ([f08d7c3](https://github.com/boltdesignsystem/bolt/commit/f08d7c3))
+
+
+
+
+
 # [2.14.0](https://github.com/boltdesignsystem/bolt/compare/v2.13.3...v2.14.0) (2020-01-06)